tooling/mpc-runtime/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
redis>=5.0.0
PyJWT[crypto]>=2.11.0
boto3>=1.34.0

